requirements.txt
discord.py==2.3.2
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.8.2
